requirements.txt
atproto>=0.0.50
requests>=2.31.0
beautifulsoup4>=4.12.0
schedule>=1.2.0
python-dotenv>=1.0.0
Pillow>=10.0.0
lxml>=4.9.0
cloudscraper>=1.2.71
